refactor(FieldTemplate): make component as proxy to render right component depending on schema
